chore: move files around
so it's easier to manage feature flag switches
